+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## 5.1.6.0 (beta) - 2025-03-XX
 * Added target acquisition summary report to Reporting display.
+* Allow the current target to continue if it can use remaining visibility time if less than project minimum.
+* Above should also help with TS Condition checks stopping the container if that same timing applies.
 * Fixed issue where targets could forget dither state.
-* Fixed issues with last flat image missing TS image file pattern variable substitution.
+* Fixed issues with last flat image missing TS image file pattern variable substitutions.
 
 ## 5.1.5.0 (beta) - 2025-03-07
 * Added explicit display of regular or provisional percent complete on exposure plans.